Update governance.md with spelling/grammar fixes.
Sorry if I did this incorrectly. I'm kind of new to Git/Github/contributing to projects, but I love OpenRefine and was interested in the governance model. I figured I would just clean up this document a little bit while I was reading through it. The changes are quite minor.