Update download.md (#85)
* Update download.md

Adding a column to the extensions table to indicate last compatible version. Saving & updating in multiple passes, so ignore for now.

* Update download.md

* Update download.md

Removed VIB-Bits "save facets and diff plugin" entry because it appears not to exist (unless someone has an old version we can link).

* Removed LMF entry

Same for LMF - removed entry unless there is a legacy download somewhere. No downloads left as far as I can see. If I understand correctly this is a standalone app (http://marmotta.apache.org/) that may or may not provide OpenRefine integrations when both are installed. Seems a bit out of scope for testing but perhaps if I have time.

"| [LMF Extension](https://bitbucket.org/srfgkmt/lmf/wiki/Google-Refine-Extension) | The Linked Media Framework is an easy-to-setup server application that bundles together some key open source projects to offer some advanced services for linked media management. The core development of this project has been moved to Apache Marmotta  |  Salzburg Research | "

* giTorto extensions

Neither are working. The Extractu plugin perhaps just needs to be compiled (which is beyond me) but the geoXtension plugin installation instructions are incomprehensible (and the reference to a provided Docker doesn't seem to be true).

* D2Refine

Not working (insofar as the easiest identifiable modification would be an added reconciliation service).

* Warning about extensions

* GOKb link

* Minor cleanup

* Version estimates based on commit dates

* Reordering based on compatibility

& putting broken ones at the bottom of the table

* More reordering

* Cleaning up language, normalizing keywords

"Exports" and "reconciles" and etc.

* Attributing

* Update download.md

* Details on FAIR

* Update download.md

* Update download.md

* Update download.md

* Update download.md

* Update download.md