Update reconciling.md (#267)
I think it is useful to include the way to access labels, aliases etc as people often struggle with this. Also I cannot find documented anywhere the way to get Commons sitelinks (adding the wiki on the end was not obvious to me!) so wanted to make sure that was added.